Support inline embed of files in the same gist
In the main Markdown file, author can embed inline any file in the
same gist by using tag `<%= files[filename] %>`

It is possible by taking advantage of the eex template with the right
bindings. In this case, we evaluate the HTML-converted Markdown
using a `files` binding with list of embed HTML partial using the file
name as key.

The embed code using GitHub's Javascript, and fall back to a `pre`
tag when `noscript`.