First attempt of revisiting pagination with cache in place
This is really tricky, at least for me. GitHub supports pagination but
we need to cache them per page(?). And we only care about gists
with Markdown file, so how it takes into account of the page?

Implemented a function to continuously fetching GitHub and cache
for `page` requested * `@per_page` (set in `env`). The result is sliced
accordingly later on. However, still has not filtered for Markdown yet.