Save rendered template for gists into HTML files
Use the file system to save rendered template into HTML files. When
request comes in, we check if we have the static file and the time-
stamp of it is newer than the start time of the cache. If not, we
render the gists from the templates, and save to HTML.

The next request will just get the static file.

Added some docs.

Temporarily remove the "Attachments" section in the template.