Split the text blocks client-side
The server should not be concerned of preparing the data for a Sir
Trevor instance on the client-side. If the client uses Sir Trevor,
then the client needs to parse the markdown data to fit what it
needs.

Therefore, the server just put the markdown inside the `textarea`.
The client takes it out, parse it for text blocks, store to JSON,
and put the stringified back into the `textarea` for Sir Trevor to
consume and render as blocks.