Avoid using tabs to have more screen estate for markdown block
Instead of having tabs taking space, switch to use only one `textarea`.
There is a new button on the block's UI, which when on hover, shows
the rendered HTML and revert back to the editor on hover out.

The textarea has its height auto adjust to the content.

Using hover makes the block flickering due to different height between
the markdown textarea and the rendered HTML. Need to fix.