Attempt to fix an infinite loop on requests
My guess is that when trying to go to `/something` where `something`
is not a valid GitHub username, GitHub still responses with 200 status,
but the error is in the response's body.

That error was not caught, and causing the fetching gists to go into an
infinite loop, which hit the rate limit.