Markdown block to render and save image's base64.
When loading data, the blocks replace any reference to a base64
image with the file's name specified in the `alt` attribute. It stores
the file's base64 into the map.

Upon receiving a file on drop, the block calculates and store the file's
base64 in a map by its filename.

On submit, it replaces any image with `src` of its name with the right
base64 content and set the `alt` attribute with the filename.

For now, the base64 content is save directly in the content of the file.

Known bug: Upon submitting, the POST redirects to the image file.