Includes Embeddable mixin.
-The embeddable mixin allows for the addition of the
embedding button that was present on the code block
to be added to other blocks.

-This has to be initialized either in the onRender function or
the loadData function.  When existing data is used to initialize
the block, then it must load the mixin there and call the
loadNonEmbedded function.  This takes care of the case
that the initializing data has a non-embedded file with the
modified name.

-When the block is created without any initial data, then the
mixin must be initialized from the onRender function, as the
loadData function will not be called.

-The image block now uses the Embeddable mixin.