Bug fixes: Deleting Gist, multiple forms, comment focus
Changed deletion to be a get on the gist handler rather
than a post.  This also cleared up some issues with the
code that occurred during our changes.

Sir Trevor library had an issue with us having multiple
Sir Trevor forms on a single page.  We commented
out a single line of the library that allowed for this
to work.

To comment the user will need to add a markdown
block through the Sir Trevor form.  This keeps it from
automatically gaining focus whenever viewing another
user's gist while signed in.